Fix the way errors are estimated in Histogram (#126)
The per-bucket StandardError accumulator needs to be corrected for the total
size of the histogram to match the way p.d.f. is normalized. Previously, it
incorrectly was corrected by adding zero samples. In fact, it needs to be
multiplied by a factor.

Also changed Histogram.Size() to Histogram.WeightsTotal(), to reflect better its
semantics.